add interface Class for vehicle
